# apps/server/swagger.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"PingController_ping","parameters":[],"responses":{"200":{"description":""}}}},"/register":{"post":{"operationId":"AuthenticationController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}}}},"/login":{"post":{"operationId":"AuthenticationController_logIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"200":{"description":""}}}},"/logout":{"post":{"operationId":"AuthenticationController_logOut","parameters":[],"responses":{"200":{"description":""}}}},"/me":{"get":{"operationId":"AuthenticationController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/reset-password":{"post":{"operationId":"AuthenticationController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"200":{"description":""}}}},"/user":{"patch":{"operationId":"UsersController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}}},"/email-confirmation/confirm":{"post":{"operationId":"EmailConfirmationController_confirm","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailDto"}}}},"responses":{"201":{"description":""}}}},"/email-confirmation/resend-confirmation-link":{"post":{"operationId":"EmailConfirmationController_resendConfirmationLink","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendEmailResponseDto"}}}}}}},"/google-authentication":{"post":{"operationId":"GoogleAuthenticationController_authenticate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenVerificationDto"}}}},"responses":{"201":{"description":""}}}},"/dashboard/summary":{"get":{"operationId":"DashboardController_getSummary","parameters":[],"responses":{"200":{"description":""}}}},"/dashboard/users":{"get":{"operationId":"DashboardController_getUsers","parameters":[{"name":"page","required":true,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Dashboard","description":"","version":"1.0","contact":{}},"tags":[{"name":"auth","description":""}],"servers":[{"url":"http://localhost:4000"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"avatar":{"type":"string"},"isEmailConfirmed":{"type":"boolean"},"isRegisteredWithGoogle":{"type":"boolean"},"lastTimeSendEmailConfirmation":{"format":"date-time","type":"string"}},"required":["id","email","name","avatar","isEmailConfirmed","isRegisteredWithGoogle","lastTimeSendEmailConfirmation"]},"UpdatePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["oldPassword","newPassword"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"ConfirmEmailDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"ResendEmailResponseDto":{"type":"object","properties":{"lastTimeSendEmailConfirmation":{"format":"date-time","type":"string"},"resendTimeConfig":{"type":"number"}},"required":["lastTimeSendEmailConfirmation","resendTimeConfig"]},"TokenVerificationDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]}}}}